requirements.txt
streamlit==1.26.0
pandas==1.5.3
requests==2.31.0
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-api-python-client==2.105.0
google-auth-httplib2==0.1.0
google-generativeai==0.1.0
python-dotenv==1.0.0
re>=2.2.1  # Regular expressions module (only required if an external package is being used)
